Edit => Save
Editing is a what a user does with a keyboard. Saving is the action the completes the edit. It is confusing to label this button "Edit." I think "Save" is the most appropriate verb here. "Apply" only works for settings and "Publish" suggests that the results will appear somewhere publicly and immediately.